Authentication

+++ app/graphql/mutations/destroy_event.rb
@@ -6,14 +6,8 @@ module Mutations
 
     def resolve(id:)
       event = Event.find(id)
-
-      begin
-        event.destroy
-        { id: }
-      rescue ActiveRecord::RecordInvalid => e
-        GraphQL::ExecutionError.new("Invalid attributes for #{e.record.class}:" \
-                                    " #{e.record.errors.full_messages.join(', ')}")
-      end
+      event.destroy if authorize_user(event)
+      { id: }
     end
   end
 end
